- Added new features to be able to handle nodes as pure structs - FasterList Clear has been renamed to FastClear. The standard Clear will clean the array like the normal List does - An engine can now receive callback for when nodes are disabled and enabled -

+++ ECS/IEnginesRoot.cs
@@ -13,6 +13,6 @@ namespace Svelto.ECS
 
         void BuildMetaEntity(int metaEntityID, EntityDescriptor ED);
 
-        void BuildEntityInGroup(int entityID, int groupID, EntityDescriptor ED);
+        void BuildEntityInGroup(short entityID, short groupID, EntityDescriptor ED);
     }
 }
